// shared/types/types.ts
export type Product = {
  createdAt: string;
  title: string;
  image: string;
  price: number;
  discount: number;
  description: string;
  countryManufacturer: string;
  material: string;
  lenseMaterial: string;
  lenseWidth: string;
  connector: string;
  rating: number;
  ratesCount: number;
  id: string;
};